+++ idl-clients-bundle/src/main/java/software/sava/idl/clients/kamino/scope/entries/PythLazer.java
@@ -10,7 +10,7 @@ public record PythLazer(int index,
                         PublicKey oracle,
                         int feedId,
                         int exponent,
-                        int confidenceFactor,
+                        long confidenceFactor,
                         Set<EmaType> emaTypes,
                         ScopeEntry refPrice) implements ReferencesEntry {
